requirements.txt
flask==3.0.0
ebooklib==0.18.1
beautifulsoup4==4.12.2
Werkzeug==3.0.1
Flask-WTF==1.2.1
Flask-Session==0.5.0
openai==1.3.5
python-dotenv==1.0.0